patient & doctor profiles feedback and responsiveness done

+++ frontend/components/FileModal.js
@@ -1,22 +1,29 @@
-import React from 'react';
-import { Modal, Button } from 'react-bootstrap';
+import React, { useState } from 'react';
+import { Modal, Button, Alert } from 'react-bootstrap';
 import { useDispatch } from 'react-redux';
 import { downloadPatientDocs } from '@/app/redux/actions/patientActions';
 import { removeDocsAction } from '@/app/redux/actions/patientActions';
 
+
 const FileModal = ({ show, onHide, filePath, isPdf, fileName , onDelete}) => {
   const dispatch = useDispatch();
-  const role = JSON.parse(localStorage.getItem('userInfo')).data.user.role;
+  const [deleteConfirm, setDeleteConfirm] = useState(false);
+  const role = JSON.parse(localStorage.getItem('userInfo'))?.data.user.role;
+
   const handleDownload = () => {
     dispatch(downloadPatientDocs(fileName));
   };
 
   const handleRemove = () => {
+    if(!deleteConfirm)
+      return;
     dispatch(removeDocsAction(fileName)).then(() => {
-      onHide();
       onDelete();
+      onHide();
     });
+    setDeleteConfirm(false);
   };
+
 
   const fileNameWithoutPathAndDate = fileName.replace(/^.*[\\\/]/, '').replace(/\d+-/g, '');
 
@@ -29,10 +36,19 @@ const FileModal = ({ show, onHide, filePath, isPdf, fileName , onDelete}) => {
           </div>
           <div className="col-md-6 d-flex justify-content-end">
             {role==='patient' && <Button className="mx-2 bg-primary" onClick={handleDownload}>Download</Button>} 
-            {role==='patient' && <Button className="mx-2 bg-dark" onClick={handleRemove}>Delete</Button>}
+            {role==='patient' && <Button className="mx-2 bg-dark" onClick={()=> {setDeleteConfirm(true);}}>Delete</Button>}
           </div>
         </Modal.Title>
       </Modal.Header>
+        <div className="row col-md-11 mx-auto mt-2">
+        {deleteConfirm && <Alert variant="danger" className='row align-items-center' >
+          <div className='col-md-8'><strong>Warning!</strong> Are you sure you want to delete this file?</div>
+          <div className="d-flex justify-content-end col-md-4">
+            <Button className="mx-2 bg-danger" onClick={()=>handleRemove()}>Confirm</Button>
+            <Button className="mx-2 bg-dark" onClick={()=>setDeleteConfirm(false)}>Cancel</Button>
+          </div>
+        </Alert>}
+        </div>
       <Modal.Body>
         {isPdf ? (
           <iframe src={filePath} title="PDF Preview" width="100%" height="500px" />
